refactor: update PositionResponse DTO

+++ src/main/java/org/marllon/caip/domains/location/dto/response/PositionResponse.java
@@ -1,4 +1,4 @@
 package org.marllon.caip.domains.location.dto.response;
 
-public record PositionResponse(Integer latitude, Integer longitude) {
+public record PositionResponse(Double latitude, Double longitude) {
 }
